refact(sharepoint-connector): extracted file filtering logic into FileFilterService
