+++ services/sharepoint-connector/src/constants/defaults.constants.ts
@@ -2,6 +2,8 @@ export const DEFAULT_PROCESSING_CONCURRENCY = 4;
 export const DEFAULT_STEP_TIMEOUT_SECONDS = 30;
 export const DEFAULT_MAX_FILE_SIZE_BYTES = 209715200; // 200MB
 
+export const DEFAULT_DOWNLOAD_LOG_INTERVAL_BYTES = 1000000; // 1MB
+
 export const DEFAULT_TOKEN_EXPIRATION_BUFFER_MS = 5 * 60 * 1000; // 5 minutes
 
 export const HTTP_STATUS_OK_MAX = 299;
